!fixing issue log
from the log we can see that the cluster.worker.id is set to 0 and this means "main process" but, it has the same PID of the cluster.worker.id 1
the server has be launched with color feature "true" so this error log seems referring to the master process and
nodejs should use only the workers as listener for connection

Error: log-color is false but the message is not a string.
Message: undefined
Error: log-color is false but the message is not a string.
Message: undefined
Error: log-color is false but the message is not a string.
Message: undefined
Error: log-color is false but the message is not a string.
Message: undefined
Error: log-color is false but the message is not a string.
Message: undefined
Error: log-color is false but the message is not a string.
Message: undefined
Error: log-color is false but the message is not a string.
Message: undefined
!string !string(0)(117) /unknown 400 !string !string !string ⚷ 207 ⟳ (0)[1311064] !string  0.2920ms !string
[data] is missing.
9fedad89 GET(171)(276) / 301 104.234.115.153 89.168.29.7 http/1.1 ⚷ 180 ⟳ (2)[1311065] 'Mozilla/5... no-referer 0.5900ms 2025-03-29T06:31:33.565Z
609a9f5a GET(171)(276) / 301 104.234.115.153 89.168.29.7 http/1.1 ⚷ 185 ⟳ (3)[1311071] 'Mozilla/5... no-referer 0.4884ms 2025-03-29T06:31:55.938Z
25badf81 GET(191)(296) /manage/account/login 301 104.234.115.153 89.168.29.7 http/1.1 ⚷ 208 ⟳ (1)[1311064] 'Mozilla/5... no-referer 0.5286ms 2025-03-29T06:32:04.080Z